+++ CHANGELOG.md
@@ -78,3 +78,4 @@
     - Marker Selection Menu
 - Fix all executable-relative path issues
 - Bring back mouse support
+- Audio Lag setting in options